RAG-Based Customer Support Assistant - Final Project/requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
chromadb>=0.5.0
pypdf>=4.2.0
sentence-transformers>=2.7.0
streamlit>=1.34.0
python-dotenv>=1.0.1
fpdf2>=2.7.0
